Update patch set 1
Patch Set 1: Code-Review-1

I'm not sure if this is the appropriate resolution - see the comments of why custom graphics support was included (charger being one of them, and note that we've got derpy charger animations at the moment).

IIRC there were some slight changes to some of the APIs in AOSP that need to be taken into account.

-1 until I can take a closer look at this tonight.

Patch-set: 1
Label: Code-Review=-1