Update patch set 4
Patch Set 4: Code-Review-1

The change on storage_list.xml is correct, but the one on init is not correct.

As you can see on source.android.com. sdcard1 is a phisical storage and this one has the next configuration.

Do not follow that there appears that is a sdcard0. you need to check the type of the storage. As I spoke with XpLoDwiLD when we setted up it for Android 4.4 both should be based on phisycal storage, not any of the others options. Also on first time the init was done as on this commit, but then I had to change it due to errors.

http://i1346.photobucket.com/albums/p690/hernanmvp10/Screenshotfrom2014-02-13195849_zps5a8a55e7.png

Patch-set: 4
Label: Code-Review=-1