Update patch set 1
Patch Set 1: Verified+1

Simply use 0/1 as brightness value. Liblights must handle every value != 0 as on. Gonna add some special config for such devices in the future. And regarding timeout. If the kernel is doing button light stuff on its own this should be disabled. Liblights is the matter nobody else is allowed to change the state of button lights

Patch-set: 1
Label: Verified=+1