Update patch set 2
Patch Set 2:

The kernel has button light controls including timeout, but they don't belong there; I don't see stock having this. With liblights and the config setting, the controls also worked fine. So the problem is that the kernel has a competing implementation and that the config setting switches on too much controls (button brightness, which doesn't work); these are not to be solved in this change.

So exit button light controls...

I had a look at the kernel mDNIe implementation and there was a mismatch between the options and defaults in the device settings app, the (cm-)kernel and stock. I've aligned that as much as I could. Will do more testing, since not all features always work.

I assume the idea behind 'mDNIe scenario' in stock is that it automatically changes based on the type of app you use? And that we don't have that, so we should let the user choose?

Patch-set: 2
Reviewer: Gerrit User 1000630 <1000630@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=0